+++ Functions/Market/Groups/groupInfo.js
@@ -1,7 +1,7 @@
 module.exports = groupInfo
 
 const axios = require('axios')
-const { link , dataSource} = require('../../../../esi.json')
+const { link , dataSource} = require('../../../esi.json')
 
 async function groupInfo(groupID) {
 let returningData;