+++ Functions/Opportunities/groups.js
@@ -1,12 +1,12 @@
-module.exports = prices
+module.exports = groups
 
 const axios = require('axios')
-const { link , dataSource} = require('../../../esi.json')
+const { link, dataSource } = require('../../esi.json')
 
-async function prices() {
+async function groups() {
 let returningData;
 
-    await axios.get(`${link}markets/prices/?datasource=${dataSource}`)
+    await axios.get(`${link}opportunities/groups/?datasource=${dataSource}`)
         .then(response => {
             returningData = Promise.resolve(response.data)
         })