+++ CHANGELOG.md
@@ -27,7 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated README.md
 - Improved error handling
 - Changed constants declared as `let`s to `const`s
-- added `getSettings` and `setSettings`
+- Added `getSettings` and `setSettings`
+- Datasource is now modifiable
+- Added Faction Warfare to Sovereignty
 
 
 <!-- LINKS -->